change tests for optimize_schedule to test scheduler.py - useful pytest pattern within
